Replace calls to TB_SAVERESTORE with custom functions
TB_SAVERESTORE can be used to save and restore the buttons on a toolbar.
However, that message has a few drawbacks:

- It requires a handle to the toolbar window to work. That's
  incompatible with how the rest of the settings are saved and loaded
  and generally makes it harder to load the settings up front (before
  any of the windows have been created).
- It only works when saving data to the registry. That meant that,
  previously, there were two completely separate paths for restoring
  toolbar buttons - one for the registry and one for the config file. By
  using custom functions to save and load toolbar buttons, it means that
  buttons are restored in the same way, regardless of whether the
  settings are being saved to the registry or to the config file.

Therefore, toolbar buttons are now saved and loaded using two custom
functions. Note that those functions have been designed to save and load
data in the same format as TB_SAVERESTORE, which ensures backwards and
forwards compatibility. Tests for loading and saving toolbar buttons
have also been added.

This also fixes an issue in which the default buttons would be
incorrectly saved to the config file if a customized main toolbar was
loaded from the registry and the setting to save to the config file was
then enabled. That's because
MainToolbarPersistentSettings::m_toolbarButtons would only be updated if
the toolbar was manually customized or reset.

If the settings were initially loaded from the registry, neither of
those conditions would be met and
MainToolbarPersistentSettings::m_toolbarButtons would contain the
default buttons.