Update how Better Enums settings values are loaded from registry
Previously, in Explorerplusplus::LoadGenericSettingsFromRegistry(),
_from_integral() was being used to convert an integer value to an enum
value. However, that method will throw an exception if the provided
value isn't a valid enum value. That meant that if the stored value was
invalid (e.g. because the saved data had been corrupted), the
application would crash on startup.

To fix that, the method used to load a Better Enums value has been
updated, with a check first being performed to determine whether the
value is valid. If the value isn't valid, the original (default) value
will be used.

Note that with this change, the previous key used to store the dark mode
status (EnableDarkMode) will no longer be read. That means that a user
updating from an older dev version will need to switch to dark mode
manually via the options dialog.