Allow navigation to search-ms: URLs
A search-ms: URL represents a search results folder. It's possible to
navigate to those folders, provided the STR_PARSE_PREFER_FOLDER_BROWSING
option is set when calling SHParseDisplayName(). Therefore, that's now
done.

Note that passing the STR_PARSE_PREFER_FOLDER_BROWSING option to
SHParseDisplayName() also appears to fix the issue described in
https://explorerplusplus.com/forum/viewtopic.php?t=3185. That is, when
navigating to a folder like the "Downloads" folder, the treeview
selection wouldn't be set.

From some investigation, it appears that's because the pidl returned by
SHParseDisplayName() would refer to an item in the root desktop folder,
but not the "Downloads" item that normally appears.

For whatever reason, passing the STR_PARSE_PREFER_FOLDER_BROWSING option
causes the pidl that's returned to refer to the filesystem folder
specifically, which means that folder is what will be selected in the
treeview.

This also fixes an issue that arose if the "Open zip files in
Explorer++" setting were disabled. In that situation, no other file
container types (e.g. .7z, .cab) would be opened in Explorer++ either.
Now, disabling that option only stops .zip files from being opened in
Explorer++. That means that opening a saved search file (.search-ms)
works consistently.