Add helper function to get menu item at point
This helper function simplifies some of the code, since it will
recursively loop through each submenu of a menu to find the menu item at
a specific point. Previously, there were a few places where a list of
menus was being stored.

Tests for a number of the MenuHelper methods have also been added.