Refactor menu view code
The PopupMenuView class has been split into two classes - MenuView and
PopupMenuView. This has been done to allow another subclass of MenuView
to be added that will represent a set of items on the main menu.

Additionally, the menu view is now passed into the menu controller,
rather than having the controller create the view. That creates a
cleaner separation between the view and the controller code that manages
it.