Add global history menu
This menu can be accessed via Go > History and will display a list of
all history items, with more recent items appearing first. Items in the
menu can be both clicked (which will open the item in the current tab)
and middle-clicked (which will open the item in a new tab), with the
ctrl and shift keys used to control exactly how an entry is opened.

Note that the menu will automatically update when the history changes.
That means that, for example, if an item is middle-clicked, the
navigation will be added to the global history and the menu will be
updated automatically.