Add TabParentItemsMenu class
This class will display the set of parent items for the current tab.
Previously, the MainToolbar class retrieved a list of parent pidls and
then passed them to ShellItemsMenu.

The advantage of the current approach is that the menu is
self-contained, which makes it easier to invoke from other locations,
and the MainToolbar no longer has to determine how to retrieve a list of
parent items. It's also consistent with how the tab history menu is
managed.