Fix shortcut handling regression
In 24828ab0fb54ee0cba9a9ee6d3de9244b579d438,
Explorerplusplus::OpenItem() was updated, with the code that handles
shortcut items being simplified.

Before that change, the type of a shortcut target (folder/non-folder)
was determined and the shortcut was opened via the shell if it was a
non-folder.

After the change, the shortcut target was executed directly. That isn't
correct however, since a shortcut can have various start options defined
(e.g. parameters, initial directory). Those options won't be applied if
the target is simply executed.

Therefore, Explorerplusplus::OpenItem() has been updated so that it
will open shortcuts to folder items directly, while passing shortcuts to
files to the shell. This matches the previous correct behaviour.

This issue was reported in
https://explorerplusplus.com/forum/viewtopic.php?t=3373.