Fix window display flag passed when invoking item
Previously, when invoking an item, the SW_SHOW display flag would be
specified. However, passing that flag meant that the window state set on
a shortcut item (i.e. normal, minimized, maximized) would be ignored.

That meant that, for example, if a shortcut to an item was set to run
maximized and the shortcut was opened from Explorer++, it would open in
the normal state instead.

To fix that, the SW_SHOWNORMAL flag is now passed instead. Using that
flag will ensure that the window state set on a shortcut item will be
obeyed when opening the shortcut.