Disable automatic quick access update in treeview by default
Previously, the quick access folder in the treeview was updated when the
appropriate directory change notification was received. As the node was
collapsed and re-expanded, the treeview scroll position would jump back
to the top of the treeview.

Because the change notification can be generated for a wide variety of
events (e.g. when a file in an unrelated directory is renamed), that
could cause the treeview scroll position to often jump around, for
seemingly no reason.

Therefore, automatic updates to the quick access folder in the treeview
are now disabled by default. The functionality can be re-enabled by
passing the "--enable-features AutomaticQuickAccessUpdates" option on
the command line. Without that option, updates to the quick access
folder will be ignored.

This functionality will ultimately need to be updated to work in a more
sophisticated way.