Remove "Save search" background context menu item
This item appears on the background context menu for a search results
folder. Invoking it in Explorer++, however, doesn't work.

The item handler will attempt to retrieve view information, using
IFolderView2 and at least one undocumented interface. Because of that,
and the fact that view information in Explorer++ doesn't correspond
precisely with the view information in Explorer anyway, this menu item
is now removed before the menu is shown.