Disable paste menu item when target requires move and item is copied
The recycle bin is an example of a folder that items can be cut and
pasted into, but not copied and pasted into. In other words, it is
possible to paste into the recycle bin, provided the item is moved,
not copied.

The CFSTR_PREFERREDDROPEFFECT format is used by the source to indicate
whether a copy or a move is preferred. If the source indicates that a
move is preferred, a paste into a folder like the recycle bin is
possible. If the source indicates that a copy is preferred, a paste
isn't possible.

Therefore, the preferred drop effect will now be taken into
consideration when determining whether or not the paste menu item should
be enabled.

Note that this change also impacts the paste shortcut menu item. If the
source includes DROPEFFECT_LINK as one of its preferred options (which
should be the case when an item is copied), the paste shortcut item will
be enabled. If DROPEFFECT_LINK isn't included, the paste shortcut item
will be disabled. This matches the behaviour of Explorer.