Fix code that checks whether hard links can be pasted
In 79e5e098aa75bd9268253f4c8149884e366ef119, the code that checks to see
whether a hard link can be pasted was updated to simply check whether
the current folder is a filesystem folder. That's not correct, though,
as there needs to be data in the appropriate format on the clipboard.

Prior to that change, there was a check to see whether data could be
pasted. That isn't correct either, though. It's not enough to know that
data can be pasted, since:

- The data needs to be in a specific format (CF_HDROP).
- The paste needs to occur in a filesystem folder.

Therefore, the code that checks to see whether a hard link can be pasted
has been updated to verify the above items.