Allow symlinks to be pasted
A new "Paste Symbolic Link" menu item has been added to the "Edit" menu.
When this menu item is selected, symlinks to any files/folders on the
clipboard will be pasted into the current directory.

Whether or not the symlinks can be created depends on a few things:

- On Windows 10 and 11, developer mode needs to be on for unprivileged
  symlink creation to work. Elevation is required otherwise.
- Prior to Windows 10, elevation is always required.

As elevation isn't currently performed, that means that if elevation is
required to create the symlink, the operation will fail (unless the user
has manually run the application elevated).