Elevate symlink paste operation when necessary
As symlink creation can require elevation, pasting symlinks to any
clipboard items can fail if the process isn't currently elevated.

Therefore, the symlink paste operation will now be retried within an
elevated Explorer++ process if the original operation fails because of
insufficient privileges. Boost Interprocess is used here, in combination
with Cereal, to communicate the results of the paste operation back to
the original process.

Specifically, boost::interprocess::managed_windows_shared_memory is used
to store the results. While it would be a little simpler to use
boost::interprocess::message_queue, that has the downside of using a
backing file that needs to be explicitly cleaned up.
boost::interprocess::managed_windows_shared_memory, on the other hand,
will be destroyed automatically once all handles are closed.