Split command line string in a custom way
Previously, CommandLineToArgvW() was used to split up the command line.
However, that function works in a way that doesn't really fit in with
how filesystem paths might be passed on the command line.

For example, if the command line was:

C:\path\to\Explorer++.exe "C:\"

the path argument would be interpreted as:

C:"

due to the way in which the backslash and closing quote character were
interpreted.

While there was some code to try and detect and handle that specific
situation, it didn't work in all cases.

For example, if the command line was:

C:\path\to\Explorer++.exe "C:\" "E:\"

the path arguments would be interpreted as a single argument:

C:" E:"

This single argument would then be transformed to:

C:" E:\

which isn't at all what would be expected.

To fix these issues, the command line string is now split using the
PEGTL library. While something like Boost Tokenizer could potentially be
used, performing custom splitting of the string means that there's full
control over how the string is split. It also means that a malformed
command line (i.e. one that's missing a closing quote or has extra data
outside of a quoted argument) can be detected.

A key difference to the splitting performed by CommandLineToArgvW() is
that backslash characters are always interpreted literally. Backslashes
are never used as escape characters, nor are there any other escape
characters. Note that that means that it's not possible to use a double
quote character within an individual argument, since there's no way of
escaping the quote character.

That should be ok, since a quote character isn't a valid filename
character and can't appear anywhere in a valid filesystem path. For
virtual items in the shell namespace, it's possible the parsing path
restrictions are more relaxed, though I'm not aware of any virtual items
that have a parsing path that includes a double quote character.

Note that this also fixes an issue with pasting symlinks in a drive
root. In that situation, the application would be invoked with a command
line like:

C:\path\to\Explorer++.exe --paste-symlinks "E:\"

As with the above cases, the destination argument in that case would be
previously be interpreted as E:". Now, it's correctly interpreted as
E:\.