Fix vcpkg triplet for Helper project in Debug-Asan configuration
Previously, the triplet wasn't set in the Debug-Asan configuration. That
meant that when building the Debug-Asan configuration on AppVeyor, the
packages would be built both for the standard triplet (when starting the
Helper project build) as well as the asan triplet (when starting the
Explorer++ project build).

There's no need to build the standard triplet in the Debug-Asan
configuration, however. The triplet now set on the Helper project in the
Debug-Asan configuration will ensure that the packages are only built
for the asan triplet.

Additionally, the --overlay-triplets option is now set for the Helper
and TestExplorer++ projects in the Debug-Asan configuration.