Fix crash that could occur when retrieving network adapter status
The function used to retrieve the status (GetNetworkAdapterColumnText)
was missing a null check.

Ultimately, it would be better to retrieve this information using the
property system. However, I don't believe the network adapter properties
are publicly defined.

Which then means that the list of available columns for a folder would
have to be dynamically retrieved and doing that would involve switching
away from the current custom columns implementation. That's a much
larger change, though, and would impact how columns are saved, loaded
and sorted.

Therefore, the only change here has been to add error checking to
GetNetworkAdapterColumnText().

This fixes #455.