Fix crash that could occur on startup
When various tab events occur (e.g. when a tab is selected), the state
of the paste button in the main toolbar is set. The code responsible for
doing that implicitly assumes that there's an active tab. While that
assumption typically holds, it may not be true during startup.

That meant that, in certain cases (e.g. the situation described in
#453), a crash could occur on startup, as there would be no active tab
at the point where the state of the paste toolbar button was being set.

To fix that, the state of the button is now only set after the
application has been initialized. This is more efficient (there's no
need to repeatedly set the state of the paste button as tabs are
restored, for example) and avoids issues where there isn't yet an active
tab.

This fixes #453.