When removing deleted item, also remove from filtered list
Previously, a crash could occur in the following situation:

1. Filtering was enabled, resulting in one or more items being removed
   from the listview.
2. One of the filtered items was deleted.
3. Filtering was disabled.

Because deleted items were only being removed from the main set of
items, but not the filtered set of items, disabling filtering at step 3
would result in an attempt to restore the non-existent item.

To fix that, an item that's deleted will now also be removed from the
list of filtered items.

This fixes #448.