Fix crash that could occur after switching from thumbnails view
When the view is changed from thumbnails view to one of the other view
modes, the thumbnails image list is destroyed. Previously, however, the
image list wasn't been removed from the listview. That meant that if the
view mode was changed from thumbnails view, to small icons, list or
details view, the thumbnails image list would be destroyed, but wouldn't
be removed from the listview (when switching to one of the other view
modes, the listview image list would be overwritten, which would mask
the issue).

The listview would then try to access the image list at various points,
which could cause a crash, like the one described in #460.

There are likely a few reasons why this issue went undetected for so
long:

- The ASAN build does correctly intercept the memory functions used to
  allocate/free the image list and attempting to directly access the
  image list once it's been destroyed does result in an error. However,
  the access in this case is occurring within a system dll, which isn't
  instrumented.
- The image list implementation attempts to detect whether an image list
  is valid, based on a 4 byte magic value (this check is performed in
  comctl32.dll!CImageListBase::IsValid). As that value is unlikely to be
  set by chance, the end result is that use of a freed image list is
  unlikely to trigger any observable issues, since the validity check is
  likely to fail.
- The issue is correctly detected when using Application Verifier.
  However, that's designed as more of a manual tool, to be used while
  the application is being debugged. Because it's designed to be used
  manually, Explorer++ is rarely run under Application Verifier.

To fix this issue, the listview image list is now reset before the image
list is destroyed, when switching from thumbnails view to another view.

This fixes #460.