Update third-party include category regex
Previously, a list of third-party include prefixes was specified.
However, that list wasn't been kept up-to-date, since it's easy to
forget to update the list when adding a third-party dependency.

Therefore, a third-party include is now considered to be an include in
angle brackets that includes at least one forward slash. The idea being
that system headers (C++ headers or Windows headers) will never include
a forward slash, so an include in angle brackets that does have a
forward slash must be a third-party header.