Correct index check in MovableModel::MoveItem()
The previous check was off by one. Because the size of the model isn't
changing, an item can be moved up to the current last index. An index
past the index of the last item needs to be adjusted.

Additionally, an index check has been added to MovableModel::AddItem().

Tests for both of these cases have been added.