Include VersionNumber.props in TestExplorer++ project
This allows the build number to be accessed from within the
TestExplorer++ project, which is useful in the VersionHelper test. This
test was previously failing on AppVeyor, due to the missing build
number.

Additionally, this simplifies how the build number is set up.
Previously, VersionNumber.props set the build number in each of the
translation projects, while the Explorer++.vcxproj file did the same
thing manually.

Now, the VersionNumber.props file is consistently used to set the build
number.