Define accelerator table in code
This is part of some broader work to make accelerators updatable. To do
that, the accelerator text on menu items will need to be dynamically
generated (rather than fixed in the .rc file). That then requires that
items that have multiple accelerators have the "default" accelerator
(the one shown on the menu entry) chosen in a deterministic way.

That isn't going to work with the .rc file, since entries there can be
rearranged incidentally, when adding or updating entries. Additionally,
I'm not sure that LoadAccelerators() makes any guarantees about the
order in which items appear in the loaded table anyway.

Fixed ordering is important, since without it, items that have multiple
accelerators could have different menu accelerator text generated in
different builds. For example, a search can be initiated using either
Ctrl+F or F3. Only one of those shortcuts should appear on the menu and
that choice should be explicit and deterministic.

By defining the accelerators in code, the "default" accelerator can be
set by the order in which accelerators are defined.

This is part of the work for issue #476.