Allow string resources to be loaded using a global instance
As there's only ever going to be a single way of loading resource
strings while the application is running, passing around the resource
instance just creates clutter.

The global instance here also makes use of a ResourceLoader interface,
which more cleanly separates the task of loading a resource from the
specifics of the implementation.