Add UniqueThreadId class
This class represents a unique ID for a thread. The advantage over
std::thread::id is that values are never reused.

At the moment, this class is only used by Runtime. Whilst it isn't too
useful there (since Runtime assumes it's being created from the UI
thread and it's not possible for the UI thread to finish without the
entire application exiting), it may be useful in other places (e.g. to
check whether a particular object is being used from the same thread).