Add App class
This class manages objects that are relevant for the application as a
whole. The class has been added to make it possible to more cleanly
separate objects that are tied to the application and objects that are
tied to a top-level window (i.e. the window managed by the
Explorerplusplus class).