Simplify processing of --application-crashed argument
Previously, a CrashedDataTuple type, with the same types as those in
CrashedData, was manually declared. Now, Boost.PFR is used to
automatically create a tuple of the right type.

A library like Boost.PFR is ultimately necessary, since C++ currently
has no reflection support, so CLI11 has no way to extract the fields
from a struct.