Call MiniDumpWriteDump() directly
Previously, dbghelp.dll was manually loaded and MiniDumpWriteDump() was
looked up using GetProcAddress(). However, there's not really any reason
to do that here. dbghelp.dll is included with Windows and while it can
make sense for applications to ship an updated version of that DLL,
that's not what's done here.