Update code used to position main window
Although a WINDOWPLACEMENT struct is currently used to save the position
of the main window, the only two pieces of information that are actually
used are the window position and the show state. That's made more clear
by the fact that CreateMainWindow() now relies only on those.

This commit also includes some minor alterations to the functional
behaviour:

- The default window position/size has been updated.
- Previously, if the main window was minimized when it was closed, it
  would be minimized the next time the application was run. Now, the
  main window will be shown in its normal position instead.
- If the show state passed to WinMain() was minimized or maximized, that
  would be used instead of the loaded show state. Now, the loaded state
  is always used, if available.
  That generally makes more sense, since the window is being restored
  from saved data, so its show state should also be the saved show
  state.