Flush clipboard in ShellBrowserImpl destructor, if necessary
Previously, ShellBrowserImpl would flush the clipboard during
application shutdown. While the clipboard only needs to be flushed when
the application exits, determining whether or not the application set
the current data on the clipboard is, in general, difficult (e.g.
because the tab that set the data might have been closed). Flushing the
clipboard within the ShellBrowserImpl destructor is the easiest option.

This also fixes an issue where the clipboard data wouldn't be retained
if a cut/copy was performed in a tab and the tab was closed before the
application was closed. Since the clipboard was only being flushed by
the ShellBrowserImpl instance during application shutdown, that wouldn't
happen if the tab had already been closed.