Disable relative include paths in Helper project
Warning C4464 (relative include path contains '..') is now enabled in
the Helper project. That then means that an attempt to include a file
using a relative path will trigger a warning, which will cause the
release build to fail. That effectively disables relative include paths
in the Helper project.

That's important, since the Helper project has no dependencies on the
other projects in the solution, so files within the project should only
be including other files within the project.