Destroy TaskbarThumbnails instance when top-level window is destroyed
TaskbarThumbnails depends on the TabContainer instance. Therefore, it's
important that the TaskbarThumbnails instance is destroyed before the
TabContainer instance.

Previously, the destruction order was wrong. TabContainer would be
destroyed when the tab control was destroyed. However, TaskbarThumbnails
is owned by the Explorerplusplus class and was destroyed when that class
was destroyed (i.e. after the TabContainer had been destroyed).

Now, the TaskbarThumbnails instance is destroyed when destruction of the
top-level window begins.