Delete application registry key when saving settings
Previously, a few different save functions were clearing sub-keys before
saving. However, it's easier to simply delete the entire application
key.

Note that this does have some implications when it comes to forwards
compatibility. That is, the keys used to save data sometimes change. The
application is generally designed to be backwards compatible, which
means that the code used to load data in a previous format is never
removed. Therefore, newer versions of the application should always be
able to load data saved by older versions.

However, forwards compatibility is a lot harder to maintain, since it's
not known what changes will be made in the future. By deleting the
application key before saving, any previous data will be cleared, which
means that running a previous version after running a newer version
isn't necessarily going to work, particularly because older versions
tended to simply assume particular keys/values were present.

However, the change here is still judged to be worthwhile since:

- Deleting the application key in a single place is simpler.
- Forwards compatibility never really worked too well, anyway. For
  example, the way in which bookmarks were saved changed in a previous
  version, with bookmarks being saved before that version being migrated
  to the newer format.
  If an older version was run after running the newer version, the
  original bookmarks might still appear, but they wouldn't incorporate
  any changes made to the bookmarks in the newer version.
  Likewise, if changes were made to the bookmarks using the older
  version, then the newer version was run again, the changes wouldn't be
  picked up, since the newer version would preferentially load bookmarks
  using the newer bookmarks key.
- If an older version was run after exclusively running a newer version,
  the older version might not work. That is, although keys saved by an
  older version might have been kept, they weren't proactively created
  by a newer version, so the keys that the older version expected to be
  present might not have been present.
- In general, there is no good way of maintaining forwards
  compatibility, so spending time attempting to partially do it isn't
  worthwhile.
- The config.xml file is overwritten every time, so older settings have
  never been persisted there.