Update way in which window data is saved
Currently, only a single top-level window is shown per-process. The aim,
however, is to be able to show multiple window in a single process. To
facilitate that, the way in which window data is saved has been updated.

Previously, the details for the single top-level window were saved with
the other generic settings. Now, the details of each window are saved
under a windows-specific registry key/XML node.

This allows the details for multiple windows to be saved and loaded,
though only a single window will be shown at the moment.