Consistently check whether IXMLDOMNode::selectSingleNode() returns S_OK
The method will return S_FALSE if the node isn't found. The trouble with
that is that S_FALSE is considered to be a success result, that is
SUCCEEDED(S_FALSE) will return true. So, attempting to proceed when the
return value is S_FALSE (or anything but S_OK) isn't going to work.