Don't fall back to default UI language
Previously, there was a small amount of code to attempt to use the
user's default language as the application language. However, that was
fairly limited, as it only set the language if there was no saved
language (i.e. if the application had never been run before).

That code has been removed, which means that an explicit language
selection will now always need to be made.