Load config data in App class
Also, note that the previous registry key used to store the dark mode
status (EnableDarkMode) is now used as a fallback again. The key was
removed as a fallback in 4dfde9fd9d4ffc86728e2c322b0b3a3c81adf410.
However, since the XML config code still has the relevant fallback, it
makes sense to use the fallback when loading from the registry as well,
for consistency.