Change points at which browser window is registered/unregistered
Previously, a browser window would be registered when the associated
class (Explorerplusplus) was created and unregistered when the class was
destroyed.

However, it doesn't make much sense to do that, since a window that's
not fully constructed isn't useful and shouldn't appear in the browser
list. Similarly, a window that's started closing also isn't useful and
shouldn't appear in the browser list.

Therefore, the window will now only be added when its construction is
complete and the window has been shown. And the window will be removed
when the window destruction process starts.