Update how browser window close/application exit are requested
When the confirmCloseTabs setting is enabled, Explorer++ will ask the
user to confirm whether or not they want to close the window, if the
window has more than one tab.

Now, a similar confirmation will also be shown if there is more than one
window shown when exiting the application.

The message shown in either case now also includes the number of
tabs/windows that will be closed.