Save settings when WM_ENDSESSION is received
A WM_ENDSESSION message indicates that the session is ending. Once that
message has been received, the application will be terminated.
Previously, that meant that if the application was running when
WM_ENDSESSION was received, it would simply exit, without saving the
settings.

Now, when that message is received, the settings are saved.