Update code used to load languages
Previously, when a language other than the default was set, Explorer++
would iterate through files in the application directory, to try and
find the DLL for the language. Now, there's a fixed mapping between
language IDs and DLL language codes and the DLL will be loaded directly.

This has the advantage of not needing to iterate through files, which is
both more efficient and simpler. It does, however, mean that if a
translation project is added, renamed or removed, the list of language
mappings will have to be manually updated. That shouldn't be too big of
an issue, since the set of translations is only occasionally updated.