Remove TestConfigFile() function
Previously, this function was called to check for the existence of the
XML config file. If the file existed, the settings would then be loaded
from the config file, rather than the registry.

However, doing that had time of check, time of use issues. That is, it's
possible that the config file could exist when the check function was
called, but no longer exist by the time it came to actually load
settings.

Therefore, XmlAppStorageFactory::MaybeCreate() is now called directly.
If that function succeeds (because the config file exists and was able
to be read), then the settings are loaded from the config file.
Otherwise, settings are loaded from the registry.