Add Explorer++.args.json
This file is used with the Smart Command Line Arguments extension
(https://github.com/MBulli/SmartCommandlineArgs). It contains a list of
command line arguments that can be passed to the application.

The advantage of this is that it makes it much easier to start the
application from Visual Studio with one or more command line arguments.
Rather than having to manually enter a command line argument when
testing, the argument can be turned on simply by checking a box.
Similarly, the argument can be turned off by unchecking the box.