Add note about bookmarks toolbar bug fix to History.txt
In version 1.4, the height of the bookmarks toolbar wasn't updated when
an item was added. While that wasn't problematic at 100% and 125% DPI
scaling, at 150%, it caused an issue where, if the toolbar was initially
empty, and a bookmark was added, nothing would appear on the toolbar.

That's because the toolbar would be 1px too small vertically, which then
meant that the toolbar control wouldn't draw anything.

This works correctly now - the toolbar height is updated whenever a
bookmark is added to the toolbar. It's likely this was fixed as a side
effect of the main font changes made in
6cf8059282310ed20a8c49be98c4bba3b6245bd1. In that commit,
UpdateRebarBandSize() was updated to also set the height of the band
(and not just its ideal width).

So, although this issue wasn't explicitly fixed on its own, its worth
mentioning in History.txt anyway.