Remove XMLSettings::AddWhiteSpaceToNode()
This function was being used to add whitespace to the config file, in
order to manually format it. However, it would be better to format the
document as a whole, in an automated way, rather than manually inserting
whitespace.

Therefore, the function has been removed. Note that, with this change,
most of the config file now appears on a single line.