Use std::chrono::steady_clock in BrowserList
When a browser window is activated, the last active time will be
updated. The list of browser windows will then be indexed by that time,
with more recently activated windows appearing first.

Previously, the last active time was represented with
std::chrono::system_clock. However, using that could have surprising
effects if the system time changes. A monotonic clock is a better
option, since it will always increase. Also, the last active times don't
need to be stored here and the epoch is irrelevant; the only thing that
matters is the relative ordering of two times. Therefore,
std::chrono::steady_clock is now used to store the last active time.