Copy width of treeview/display window when creating a new window
When creating a new window, the width of the treeview/display window
will now be copied from the existing window. Previously, the values were
default initialized to 0, which meant that the widths would be set to
the minimum size.

The display window height is also set in the same way.