Fix rebar layout issue
Previously, when handling layout updates during startup, the following
sequence of events would occur:

1. A WM_SIZE message for the main window would be handled.
2. Within OnSize(), the size of the main rebar would be updated.
3. If the size of the rebar wasn't the default size (e.g. because two
   bands had been moved onto the same line), a RBN_HEIGHTCHANGE message
   would be dispatched. That message would result in another call to
   OnSize(). That call would be synchronous, meaning it would run within
   the middle of the original OnSize() call.
   Within the nested OnSize() call, the controls would be positioned
   correctly. However, once control returned to the initial OnSize()
   call, the original (and outdated) rebar height would be used to lay
   out the other controls.
   That meant that if two of the bands were on the same line, the
   controls below the main rebar wouldn't be positioned correctly.

The core issue here (the nested OnSize() calls) has been present for a
while, but wasn't noticeable before
a54c41b2056d26908cd955e49b251a02e0d5317c. In that commit, the rebar
sizing was moved to the top of OnSize(). Because the sizing was
originally performed at the very end of OnSize(), the issue described
above wouldn't have been apparent. That's because the nested OnSize()
call would run once the initial OnSize() call was effectively finished.

However, the issue wasn't visible to the user until
e5797129448f48691d0da1107ddfcb717bbf85d7. In that commit, the size of
the main window was set during the window creation. Previously, the size
of the window was set using SetWindowPlacement(). The call to
SetWindowPlacement() resulted in a final call to OnSize(), which set the
position of the controls correctly.

Without the call to SetWindowPlacement(), the situation described at the
top of this message would take place, resulting in the controls not
being positioned correctly.

To fix this issue, a layout update will now be scheduled when a
RBN_HEIGHTCHANGE message is received, rather than being run immediately.
Additionally, a re-entrancy check has also been added to OnSize().